Add styling to fade out long descriptions
Short-term implementation, as there are some limitations to how
this component is/was structured. Lab descriptions are often long,
and so we only display the full length description in a modal.
Previous implementation used javascript to truncate the lab body
string, which could cause HTML breaks. Here I hide overflowing
content and apply a fade-out to indicate there is more to read.