Fix NavBar component
The NavBar component previously acted as a container for three buttons,
and the other two buttons in the nav bar were implemented in each page.
This commit introduces slots with default content, so now the nav bar
contains all the nav elements while providing flexibility when content
needs to change per page.