Adjust dropdown styling to match :focus styling to default
When the account menu toggle is focused, it maintains the active
styling of the dropdown menu even if the menu is closed. This can
be confusing to users, so we adjust the color styling to match
the default styling of the button but maintain the focus outline
for accessibility. I had to remove the CSS scoping to keep the
classnames the same so I could target the button (Vue-Bootstrap
adds classes to the component that don't take scoping into account).